+++ mpp/src/ModelStream.cpp
@@ -64,7 +64,7 @@ namespace mpp
 		// Do our channels have the same data type as the stream, and do they map
 		// in the same order?
 		int vertexOffset = 0;
-		for (int i = 0; i < bufferSpec.getNumAttributes(); ++i)
+		for (size_t i = 0; i < bufferSpec.getNumAttributes(); ++i)
 		{
 			auto const& attrib = bufferSpec.getAttribute(i);
 			auto const& stream = componentStreams.at(attrib.component);
@@ -113,7 +113,7 @@ namespace mpp
 
 		for (int i = 0; i < vertexCount; ++i)
 		{
-			for (int j = 0; j < bufferSpec.getNumAttributes(); ++j)
+			for (size_t j = 0; j < bufferSpec.getNumAttributes(); ++j)
 			{
 				auto const& attrib = bufferSpec.getAttribute(j);
 				auto const& stream = componentStreams.at(attrib.component);
@@ -198,14 +198,14 @@ namespace mpp
 			}
 
 			
-			for (int j = 0; j < meshSpec.getNumVertexBufferAttributeLayouts(); ++j)
+			for (size_t j = 0; j < meshSpec.getNumVertexBufferAttributeLayouts(); ++j)
 			{
 				auto const& bufferSpec = meshSpec.getVertexBufferAttributeLayout(j);
 
 				// Acquire the vertex streams needed and work out the stride.
 				int vertexStride = 0;
 				map<Vertex::Component, VertexDataStreamDefinition> componentStreams;
-				for (int k = 0; k < bufferSpec.getNumAttributes(); ++k)
+				for (size_t k = 0; k < bufferSpec.getNumAttributes(); ++k)
 				{
 					auto const& attrib = bufferSpec.getAttribute(k);
