+++ mpp/src/SphereModelStream.cpp
@@ -74,11 +74,11 @@ namespace mpp
 		mMeshDataDefinition.vertexData.resize(bufferSize);
 
 		// Set vertex data
-		for (int i = 0; i < meshSpec.getNumVertexBufferAttributeLayouts(); ++i)
+		for (size_t i = 0; i < meshSpec.getNumVertexBufferAttributeLayouts(); ++i)
 		{
 			auto const& layout = meshSpec.getVertexBufferAttributeLayout(i);
 
-			for (int j = 0; j < layout.getNumAttributes(); ++j)
+			for (size_t j = 0; j < layout.getNumAttributes(); ++j)
 			{
 				auto const& attrib = layout.getAttribute(j);
