+++ mpp/src/TriangleBatch.cpp
@@ -122,7 +122,7 @@ namespace mpp
 			primitiveCount,
 			mesh::VertexBufferStorageType::Dynamic);
 
-		for (int i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
+		for (size_t i = 0; i < mSpecification.getNumVertexBufferAttributeLayouts(); ++i)
 		{
 			auto const& layout = mSpecification.getVertexBufferAttributeLayout(i);
 			createVertexBuffer(i, mesh, vertexCount, layout.isStatic());
